I created a new package in use_case for data access interfaces. Then, I renamed ethan's DataAccessInterface to GameDataAccessInterface and I removed the getPromptGenerator() method so that it only has save(Game) and getGame(). This class can be used by use cases that don't need to generate stories or prompts etc. I made another class called PromptGameDataAccessInterface that extends GameDataAccessInterface and it contains a method getPromptGenerator() and a method save(PromptGenerator). Since it extends the other class it has the game related methods too. This is the interface my vote out use case and Ethan's kill villager use case should be using.
