+++ src/data_access/DummyChatGPTAPI.java
@@ -1,6 +1,6 @@
 package data_access;
 
-import use_case.ChatAPIAccessInterface;
+import use_case.data_access.ChatAPIAccessInterface;
 
 public class DummyChatGPTAPI implements ChatAPIAccessInterface {
     @Override
