+++ src/use_case/data_access_interface/ChatAPIAccessInterface.java
@@ -1,4 +1,4 @@
-package use_case;
+package use_case.data_access_interface;
 
 // Might be in the wrong place
 public interface ChatAPIAccessInterface {
