+++ src/use_case/data_access_interface/GameDataAccessInterface.java
@@ -1,0 +1,8 @@
+package use_case.data_access_interface;
+
+import entity.Game;
+
+public interface GameDataAccessInterface {
+    void save(Game game);
+    Game getGame();
+}
