+++ src/use_case/data_access_interface/PromptGameDataAccessInterface.java
@@ -1,12 +1,9 @@
-package use_case;
+package use_case.data_access_interface;
 
-import entity.Game;
 import entity.PromptGenerator;
 
-public interface DataAccessInterface {
-    void save(Game game);
-    Game getGame();
-
+public interface PromptGameDataAccessInterface extends GameDataAccessInterface {
     PromptGenerator getPromptGenerator();
     // getPromptGenerator in DAO should be pulling the ConversationHistory to then construct a PromptGenerator off of
+    void save(PromptGenerator promptGenerator);
 }
